+++ scripts/codegen/02-constants.go
@@ -1,16 +1,45 @@
 package main
 
-import "regexp"
+import (
+	"regexp"
+	"strings"
+)
 
 const (
 	CLDR_VERSION      = "31.0.1"
 	RAW_DIR           = "./raw-data"
+	CLDR_DIR          = "./data/cldr/date_translation_data"
 	SUPPLEMENTARY_DIR = "./data/supplementary/date_translation_data"
 )
 
 var (
-	rxLocale        = regexp.MustCompile(`-[A-Z0-9]+$`)
-	rxLocaleCleaner = regexp.MustCompile(`-\w+`)
+	apostropheLookAlikeChars = []string{
+		`\x{2019}`, // right single quotation mark
+		`\x{02bc}`, // modifier letter apostrophe
+		`\x{02bb}`, // modifier letter turned comma
+		`\x{055a}`, // armenian apostrophe
+		`\x{a78c}`, // latin small letter saltillo
+		`\x{2032}`, // prime
+		`\x{2035}`, // reversed prime
+		`\x{02b9}`, // modifier letter prime
+		`\x{ff07}`, // fullwidth apostrophe
+	}
+
+	numberChars = []string{
+		`0-9`,
+		`\x{0660}-\x{0669}`, // arabic
+		`\x{0f20}-\x{0f29}`, // tibetan
+	}
+
+	rxLocale              = regexp.MustCompile(`-[A-Z0-9]+$`)
+	rxLocaleCleaner       = regexp.MustCompile(`-\w+`)
+	rxDateOrderPattern    = regexp.MustCompile(`([DMY])+\x{200f}*[-/. \t]*([DMY])+\x{200f}*[-/. \t]*([DMY])+`)
+	rxNotRelativePattern  = regexp.MustCompile(`[\+\-]\s*\{0\}`)
+	rxPlusMinus           = regexp.MustCompile(`[\+\-]`)
+	rxDefaultMonthPattern = regexp.MustCompile(`^m?[` + strings.Join(numberChars, "") + `]+$`)
+	rxSanitizeAposthrope  = regexp.MustCompile(strings.Join(apostropheLookAlikeChars, "|"))
+	rxAmPmPattern         = regexp.MustCompile(`^\s*[AaPp]\s*\.?\s*[Mm]\s*\.?\s*$`)
+	rxParenthesisPattern  = regexp.MustCompile(`[\(\)]`)
 
 	// Languages with insufficient translation data are excluded
 	excludedLanguages = map[string]struct{}{